docs: report to week03.md
